requirements.txt
PyQt5>=5.15
sounddevice>=0.4
numpy>=1.20
scipy>=1.5
PyYAML>=6.0
